requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.2
Flask-Mail==0.10.0
psycopg[binary]
python-dotenv==1.0.1
bcrypt==4.2.1
pytest==8.3.4
email-validator==2.1.0.post1